Add a separate Logger.hs module
