+++ app/Icepeak.hs
@@ -3,13 +3,13 @@ module Main where
 
 import Control.Monad (void)
 import Control.Concurrent.Async
-import Control.Concurrent.STM (atomically)
-import Control.Concurrent.STM.TBQueue (readTBQueue)
+import Prelude hiding (log)
 
 import qualified System.Posix.Signals as Signals
 import qualified Control.Concurrent.Async as Async
 
-import Core (Core, coreLogRecords)
+import Core (Core (..))
+import Logger (log, processLogRecords)
 
 import qualified Core
 import qualified HttpServer
@@ -23,7 +23,7 @@ installHandlers core serverThread =
     handle = do
       Core.postQuit core
       Async.cancel serverThread
-      Core.log "\nTermination sequence initiated ..." core
+      log "\nTermination sequence initiated ..." (coreLogRecords core)
     handler = Signals.CatchOnce handle
     blockSignals = Nothing
     installHandler signal = Signals.installHandler signal handler blockSignals
@@ -39,23 +39,10 @@ main = do
   pops <- Async.async $ Core.processOps core
   upds <- Async.async $ WebsocketServer.processUpdates core
   serv <- Async.async $ Server.runServer wsServer httpServer
-  logger <- Async.async $ processLogRecords core
+  logger <- Async.async $ processLogRecords (coreLogRecords core)
   installHandlers core serv
-  Core.log "System online. ** robot sounds **" core
+  log "System online. ** robot sounds **" (coreLogRecords core)
   void $ Async.wait pops
   void $ Async.wait upds
   void $ Async.wait serv
   void $ Async.wait logger
-
-
-processLogRecords :: Core -> IO ()
-processLogRecords core = go
-  where
-    go = do
-      maybeLogRecord <- atomically $ readTBQueue (coreLogRecords core)
-      case maybeLogRecord of
-        Just logRecord -> do
-          putStrLn $ show logRecord
-          go
-        -- stop the loop
-        Nothing -> pure ()
